+++ cmd/halite/main.go
@@ -29,7 +29,7 @@ import (
 	"github.com/edlitmus/halite/internal/sls"
 )
 
-const version = "0.5.0"
+const version = "0.6.0"
 
 func main() {
 	if len(os.Args) < 2 {